+++ packages/polaris-viz/src/utilities/getLoadAnimationDelay.ts
@@ -1,0 +1,5 @@
+import {LOAD_ANIMATION_DURATION} from '@shopify/polaris-viz-core';
+
+export function getLoadAnimationDelay(index, dataLength) {
+  return index * (LOAD_ANIMATION_DURATION / dataLength);
+}